# openapi.json
{"openapi":"3.0.0","paths":{"/api/health":{"get":{"operationId":"AppController_getHealth","parameters":[],"responses":{"200":{"description":""}},"summary":"Health check endpoint","tags":["health"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["auth"]}},"/api/users/me":{"get":{"operationId":"UsersController_getCurrentUser","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["users"]},"patch":{"operationId":"UsersController_updateCurrentUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["users"]}},"/api/users":{"get":{"operationId":"UsersController_listUsers","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["users"]}},"/api/users/{id}/role":{"patch":{"operationId":"UsersController_updateUserRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRoleDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["users"]}},"/api/boards":{"get":{"operationId":"BoardsController_listBoards","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"post":{"operationId":"BoardsController_createBoard","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBoardDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}":{"get":{"operationId":"BoardsController_getBoard","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"patch":{"operationId":"BoardsController_updateBoard","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBoardDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"delete":{"operationId":"BoardsController_deleteBoard","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/members":{"post":{"operationId":"BoardsController_addMember","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddBoardMemberDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/members/{memberId}":{"delete":{"operationId":"BoardsController_removeMember","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"memberId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/columns":{"post":{"operationId":"BoardsController_createColumn","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateColumnDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/columns/{columnId}":{"patch":{"operationId":"BoardsController_updateColumn","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"columnId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateColumnDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"delete":{"operationId":"BoardsController_deleteColumn","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"columnId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/columns/reorder":{"patch":{"operationId":"BoardsController_reorderColumns","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReorderColumnsDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/columns/{columnId}/cards":{"post":{"operationId":"BoardsController_createCard","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"columnId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCardDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}":{"patch":{"operationId":"BoardsController_updateCard","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"delete":{"operationId":"BoardsController_deleteCard","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/move":{"patch":{"operationId":"BoardsController_moveCard","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoveCardDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/assignee":{"patch":{"operationId":"BoardsController_updateCardAssignee","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardAssigneeDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/watchers":{"patch":{"operationId":"BoardsController_updateCardWatchers","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardWatchersDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/labels":{"patch":{"operationId":"BoardsController_updateCardLabels","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardLabelsDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/checklist":{"post":{"operationId":"BoardsController_createChecklistItem","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChecklistItemDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/checklist/{itemId}":{"patch":{"operationId":"BoardsController_updateChecklistItem","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"itemId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateChecklistItemDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"delete":{"operationId":"BoardsController_deleteChecklistItem","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"itemId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/comments":{"post":{"operationId":"BoardsController_addComment","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddCommentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/comments/{commentId}":{"delete":{"operationId":"BoardsController_deleteComment","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"commentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/attachments":{"post":{"operationId":"BoardsController_addAttachment","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddAttachmentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/attachments/{attachmentId}":{"delete":{"operationId":"BoardsController_deleteAttachment","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"attachmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/cards/{cardId}/sprint":{"patch":{"operationId":"BoardsController_updateCardSprint","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardSprintDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/labels":{"post":{"operationId":"BoardsController_createLabel","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLabelDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/labels/{labelId}":{"patch":{"operationId":"BoardsController_updateLabel","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"labelId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLabelDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"delete":{"operationId":"BoardsController_deleteLabel","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"labelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/sprints":{"post":{"operationId":"BoardsController_createSprint","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSprintDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/sprints/{sprintId}":{"patch":{"operationId":"BoardsController_updateSprint","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"sprintId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSprintDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]},"delete":{"operationId":"BoardsController_deleteSprint","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"sprintId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}},"/api/boards/{boardId}/sprints/{sprintId}/status":{"patch":{"operationId":"BoardsController_updateSprintStatus","parameters":[{"name":"boardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"sprintId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSprintStatusDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["boards"]}}},"info":{"title":"Workboard API","description":"Hybrid Trello / Jira board management API","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","minLength":6},"displayName":{"type":"string","example":"Jane Doe"},"role":{"type":"string","enum":["ADMIN","USER"]}},"required":["email","password","displayName"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","minLength":6}},"required":["email","password"]},"UpdateUserDto":{"type":"object","properties":{"displayName":{"type":"string","example":"Jane Doe"},"password":{"type":"string","minLength":6}}},"UpdateUserRoleDto":{"type":"object","properties":{"role":{"type":"string","enum":["ADMIN","USER"]}},"required":["role"]},"CreateBoardDto":{"type":"object","properties":{"name":{"type":"string","example":"Product Roadmap"},"description":{"type":"string","example":"Planning the next major release"}},"required":["name"]},"UpdateBoardDto":{"type":"object","properties":{"name":{"type":"string","example":"Release Readiness"},"description":{"type":"string","example":"Updated description for stakeholders"}}},"AddBoardMemberDto":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"}},"required":["userId"]},"CreateColumnDto":{"type":"object","properties":{"title":{"type":"string","example":"To Do"},"position":{"type":"number","minimum":0}},"required":["title"]},"UpdateColumnDto":{"type":"object","properties":{"title":{"type":"string","example":"In Progress"},"position":{"type":"number","minimum":0}}},"ReorderColumnsDto":{"type":"object","properties":{"columnOrder":{"description":"Ordered list of column identifiers","type":"array","items":{"type":"string"}}},"required":["columnOrder"]},"CreateCardDto":{"type":"object","properties":{"title":{"type":"string","example":"Implement authentication flow"},"description":{"type":"string","example":"Handle login, registration, and forgot password"},"priority":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"type":{"type":"string","enum":["TASK","BUG","STORY","EPIC"]},"dueDate":{"type":"string","format":"date-time"},"storyPoints":{"type":"number","minimum":0},"assigneeId":{"type":"string","format":"uuid"},"reporterId":{"type":"string","format":"uuid"},"labelIds":{"description":"Associated label identifiers","type":"array","items":{"type":"string"}}},"required":["title"]},"UpdateCardDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"priority":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"type":{"type":"string","enum":["TASK","BUG","STORY","EPIC"]},"dueDate":{"type":"string","format":"date-time"},"storyPoints":{"type":"number","minimum":0},"assigneeId":{"type":"string","format":"uuid"},"reporterId":{"type":"string","format":"uuid"},"labelIds":{"description":"Associated label identifiers","type":"array","items":{"type":"string"}}}},"MoveCardDto":{"type":"object","properties":{"targetColumnId":{"type":"string","format":"uuid"},"targetPosition":{"type":"number","minimum":0,"description":"New zero-based index within the target column"}},"required":["targetColumnId"]},"UpdateCardAssigneeDto":{"type":"object","properties":{"assigneeId":{"type":"string","format":"uuid","description":"Leave empty to unassign"}}},"UpdateCardWatchersDto":{"type":"object","properties":{"watcherIds":{"description":"Complete list of watcher user ids","type":"array","items":{"type":"string"}}},"required":["watcherIds"]},"UpdateCardLabelsDto":{"type":"object","properties":{"labelIds":{"description":"Complete list of label ids applied to the card","type":"array","items":{"type":"string"}}},"required":["labelIds"]},"CreateChecklistItemDto":{"type":"object","properties":{"content":{"type":"string","example":"Write unit tests"}},"required":["content"]},"UpdateChecklistItemDto":{"type":"object","properties":{"content":{"type":"string","example":"Write integration tests"},"completed":{"type":"boolean"},"position":{"type":"number","minimum":0}}},"AddCommentDto":{"type":"object","properties":{"content":{"type":"string","example":"Remember to sync with the QA team on staging."}},"required":["content"]},"AddAttachmentDto":{"type":"object","properties":{"name":{"type":"string","example":"UI Mockup"},"url":{"type":"string","example":"https://example.com/mockup.png"},"mimeType":{"type":"string","example":"image/png"}},"required":["name","url"]},"UpdateCardSprintDto":{"type":"object","properties":{"sprintId":{"type":"string","format":"uuid","description":"Sprint identifier to link the card to. Omit to remove from sprint."}}},"CreateLabelDto":{"type":"object","properties":{"name":{"type":"string","example":"Backend"},"color":{"type":"string","example":"#ffb020"}},"required":["name","color"]},"UpdateLabelDto":{"type":"object","properties":{"name":{"type":"string","example":"Frontend"},"color":{"type":"string","example":"#34a853"}}},"CreateSprintDto":{"type":"object","properties":{"name":{"type":"string","example":"Sprint 12"},"goal":{"type":"string","example":"Finalize payment integration"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"}},"required":["name"]},"UpdateSprintDto":{"type":"object","properties":{"name":{"type":"string","example":"Sprint 12 - Extended"},"goal":{"type":"string","example":"Expanded to cover regression testing"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"}}},"UpdateSprintStatusDto":{"type":"object","properties":{"status":{"type":"string","enum":["PLANNED","ACTIVE","COMPLETED","CANCELLED"]}},"required":["status"]}}}}